refactor: rename tableview
